+++ DataRepo/tests/test_models.py
@@ -14,7 +14,6 @@ from DataRepo.models import (
     Compound,
     ElementLabel,
     Infusate,
-    InfusateTracer,
     MSRun,
     PeakData,
     PeakDataLabel,
@@ -26,8 +25,6 @@ from DataRepo.models import (
     Sample,
     Study,
     Tissue,
-    Tracer,
-    TracerLabel,
     buffer_size,
 )
 from DataRepo.models.hier_cached_model import set_cache
@@ -45,9 +42,7 @@ from DataRepo.utils import (
     IsotopeObservationParsingError,
     IsotopeParsingError,
     MissingCompounds,
-    NoSamplesError,
     SampleTableLoader,
-    UnskippedBlanksError,
     leaderboard_data,
     parse_infusate_name,
     parse_tracer_concentrations,
@@ -1922,598 +1917,6 @@ class AnimalAndSampleLoadingTests(TracebaseTestCase):
             dupes,
         )
         self.assertEqual([0, 1], rows)
-
-
-@override_settings(CACHES=settings.TEST_CACHES)
-class AccuCorDataLoadingTests(TracebaseTestCase):
-    @classmethod
-    def setUpTestData(cls):
-        call_command(
-            "load_study",
-            "DataRepo/example_data/small_dataset/small_obob_study_prerequisites.yaml",
-        )
-
-        call_command(
-            "load_animals_and_samples",
-            animal_and_sample_table_filename=(
-                "DataRepo/example_data/small_dataset/"
-                "small_obob_animal_and_sample_table.xlsx"
-            ),
-        )
-
-        super().setUpTestData()
-
-    def test_accucor_load_blank_fail(self):
-        with self.assertRaises(AggregatedErrors, msg="1 samples are missing.") as ar:
-            call_command(
-                "load_accucor_msruns",
-                accucor_file="DataRepo/example_data/small_dataset/small_obob_maven_6eaas_inf_blank_sample.xlsx",
-                protocol="Default",
-                date="2021-04-29",
-                researcher="Michael Neinast",
-                new_researcher=True,
-            )
-        aes = ar.exception
-        self.assertEqual(1, len(aes.exceptions))
-        self.assertTrue(isinstance(aes.exceptions[0], UnskippedBlanksError))
-
-    def test_accucor_load_blank_skip(self):
-        call_command(
-            "load_accucor_msruns",
-            accucor_file="DataRepo/example_data/small_dataset/small_obob_maven_6eaas_inf_blank_sample.xlsx",
-            skip_samples=("blank"),
-            protocol="Default",
-            date="2021-04-29",
-            researcher="Michael Neinast",
-            new_researcher=True,
-        )
-        SAMPLES_COUNT = 14
-        PEAKDATA_ROWS = 11
-        MEASURED_COMPOUNDS_COUNT = 2  # Glucose and lactate
-
-        self.assertEqual(
-            PeakGroup.objects.count(), MEASURED_COMPOUNDS_COUNT * SAMPLES_COUNT
-        )
-        self.assertEqual(PeakData.objects.all().count(), PEAKDATA_ROWS * SAMPLES_COUNT)
-
-    def test_accucor_load_sample_prefix(self):
-        call_command(
-            "load_accucor_msruns",
-            accucor_file="DataRepo/example_data/small_dataset/small_obob_maven_6eaas_inf_req_prefix.xlsx",
-            sample_name_prefix="PREFIX_",
-            skip_samples=("blank"),
-            protocol="Default",
-            date="2021-04-29",
-            researcher="Michael Neinast",
-            new_researcher=True,
-        )
-        SAMPLES_COUNT = 1
-        PEAKDATA_ROWS = 11
-        MEASURED_COMPOUNDS_COUNT = 2  # Glucose and lactate
-
-        self.assertEqual(
-            PeakGroup.objects.count(), MEASURED_COMPOUNDS_COUNT * SAMPLES_COUNT
-        )
-        self.assertEqual(PeakData.objects.all().count(), PEAKDATA_ROWS * SAMPLES_COUNT)
-
-    def test_accucor_load_sample_prefix_missing(self):
-        with self.assertRaises(AggregatedErrors, msg="1 samples are missing.") as ar:
-            call_command(
-                "load_accucor_msruns",
-                accucor_file="DataRepo/example_data/small_dataset/small_obob_maven_6eaas_inf_req_prefix.xlsx",
-                skip_samples=("blank"),
-                protocol="Default",
-                date="2021-04-29",
-                researcher="Michael Neinast",
-                new_researcher=True,
-            )
-        aes = ar.exception
-        nl = "\n"
-        self.assertEqual(
-            1,
-            len(aes.exceptions),
-            msg=(
-                f"Should be 1 error (NoSamplesError), but there were {len(aes.exceptions)} "
-                f"errors:{nl}{nl.join(list(map(lambda s: str(s), aes.exceptions)))}"
-            ),
-        )
-        self.assertTrue(isinstance(aes.exceptions[0], NoSamplesError))
-
-    def test_accucor_load_in_debug(self):
-
-        pre_load_counts = self.get_record_counts()
-        pre_load_maintained_values = get_all_maintained_field_values("DataRepo.models")
-        self.assertGreater(
-            len(pre_load_maintained_values.keys()),
-            0,
-            msg="Ensure there is data in the database before the test",
-        )
-        self.assertEqual(0, buffer_size(), msg="Autoupdate buffer is empty to start.")
-
-        with self.assertRaises(DryRun):
-            call_command(
-                "load_accucor_msruns",
-                accucor_file="DataRepo/example_data/small_dataset/small_obob_maven_6eaas_inf_blank_sample.xlsx",
-                skip_samples=("blank"),
-                protocol="Default",
-                date="2021-04-29",
-                researcher="Michael Neinast",
-                new_researcher=True,
-                debug=True,
-            )
-
-        post_load_maintained_values = get_all_maintained_field_values("DataRepo.models")
-        post_load_counts = self.get_record_counts()
-
-        self.assertEqual(
-            pre_load_counts,
-            post_load_counts,
-            msg="DryRun mode doesn't change any table's record count.",
-        )
-        self.assertEqual(
-            pre_load_maintained_values,
-            post_load_maintained_values,
-            msg="DryRun mode doesn't autoupdate.",
-        )
-        self.assertEqual(
-            0, buffer_size(), msg="DryRun mode doesn't leave buffered autoupdates."
-        )
-
-    def test_record_missing_compound(self):
-        adl = AccuCorDataLoader(None, None, "1972-11-24", "", "", "")
-        adl.record_missing_compound("new compound", "C1H4", 9)
-        self.assertEqual(
-            {
-                "new compound": {
-                    "formula": ["C1H4"],
-                    "rownums": [11],
-                }
-            },
-            adl.missing_compounds,
-        )
-
-
-@override_settings(CACHES=settings.TEST_CACHES)
-class IsoCorrDataLoadingTests(TracebaseTestCase):
-    @classmethod
-    def setUpTestData(cls):
-        call_command(
-            "load_study",
-            "DataRepo/example_data/protocols/loading.yaml",
-            verbosity=2,
-        )
-        call_command(
-            "load_study",
-            "DataRepo/example_data/tissues/loading.yaml",
-            verbosity=2,
-        )
-        call_command(
-            "load_compounds",
-            compounds="DataRepo/example_data/consolidated_tracebase_compound_list.tsv",
-            verbosity=2,
-        )
-
-        call_command(
-            "load_animals_and_samples",
-            animal_and_sample_table_filename=(
-                "DataRepo/example_data/AsaelR_13C-Valine+PI3Ki_flank-KPC_2021-12_isocorr_CN-corrected/"
-                "TraceBase Animal and Sample Table Templates_AR.xlsx"
-            ),
-            skip_researcher_check=True,
-        )
-
-        super().setUpTestData()
-
-    def load_multitracer_data(self):
-        call_command(
-            "load_animals_and_samples",
-            animal_and_sample_table_filename=(
-                "DataRepo/example_data/obob_fasted_ace_glycerol_3hb_citrate_eaa_fa_multiple_tracers/"
-                "animal_sample_table.xlsx"
-            ),
-            skip_researcher_check=True,
-        )
-
-        num_samples = 120
-        num_infusates = 2
-        num_infusatetracers = 3
-        num_tracers = 6
-        num_tracerlabels = 12
-
-        return (
-            num_samples,
-            num_infusates,
-            num_infusatetracers,
-            num_tracers,
-            num_tracerlabels,
-        )
-
-    def load_multilabel_data(self):
-        call_command(
-            "load_animals_and_samples",
-            animal_and_sample_table_filename=(
-                "DataRepo/example_data/obob_fasted_glc_lac_gln_ala_multiple_labels/animal_sample_table.xlsx"
-            ),
-            skip_researcher_check=True,
-        )
-
-        num_samples = 156
-        num_infusates = 2
-        num_infusatetracers = 2
-        num_tracers = 2
-        num_tracerlabels = 3
-
-        return (
-            num_samples,
-            num_infusates,
-            num_infusatetracers,
-            num_tracers,
-            num_tracerlabels,
-        )
-
-    def test_singly_labeled_isocorr_load(self):
-        pre_pg_load_count = PeakGroup.objects.count()
-        call_command(
-            "load_accucor_msruns",
-            accucor_file="DataRepo/example_data/AsaelR_13C-Valine+PI3Ki_flank-KPC_2021-12_isocorr_CN-corrected/"
-            "Serum results_cor.csv",
-            skip_samples=(
-                "Blank01",
-                "Blank02",
-                "Blank03",
-                "Blank04",
-            ),
-            protocol="Default",
-            date="2021-04-29",
-            researcher="Michael Neinast",
-            new_researcher=True,
-            isocorr_format=True,
-        )
-        post_pg_load_count = PeakGroup.objects.count()
-        # The number of samples in the isocorr csv file (not the samples file)
-        SAMPLES_COUNT = 19
-        PEAKDATA_ROWS = 24
-        MEASURED_COMPOUNDS_COUNT = 6
-
-        self.assertEqual(
-            post_pg_load_count - pre_pg_load_count,
-            MEASURED_COMPOUNDS_COUNT * SAMPLES_COUNT,
-            msg=f"PeakGroup record count should be the number of compounds [{MEASURED_COMPOUNDS_COUNT}] times the "
-            f"number of samples [{SAMPLES_COUNT}] = [{MEASURED_COMPOUNDS_COUNT * SAMPLES_COUNT}].",
-        )
-        self.assertEqual(
-            PeakData.objects.all().count(),
-            PEAKDATA_ROWS * SAMPLES_COUNT,
-            msg=f"PeakData record count should be the number of peakdata rows [{PEAKDATA_ROWS}] times the number of "
-            f"samples [{SAMPLES_COUNT}] = [{PEAKDATA_ROWS * SAMPLES_COUNT}].",
-        )
-
-    def test_singly_labeled_isocorr_missing_flag_error(self):
-        """
-        Test to make sure --isocorr-format is suggested when not supplied
-        """
-        with self.assertRaises(AggregatedErrors) as ar:
-            call_command(
-                "load_accucor_msruns",
-                accucor_file="DataRepo/example_data/AsaelR_13C-Valine+PI3Ki_flank-KPC_2021-12_isocorr_CN-corrected/"
-                "Serum results_cor.csv",
-                skip_samples=(
-                    "Blank01",
-                    "Blank02",
-                    "Blank03",
-                    "Blank04",
-                ),
-                protocol="Default",
-                date="2021-04-29",
-                researcher="Michael Neinast",
-                new_researcher=True,
-            )
-        aes = ar.exception
-        self.assertEqual(1, len(aes.exceptions))
-        self.assertIn("--isocorr-format", str(aes.exceptions[0]))
-
-    def test_multitracer_sample_table_load(self):
-        num_samples = 120
-        num_infusates = 2
-        num_infusatetracers = 9
-        num_tracers = 9
-        num_tracerlabels = 12
-
-        pre_load_sample_count = Sample.objects.count()
-        pre_load_infusate_count = Infusate.objects.count()
-        pre_load_infusatetracer_count = InfusateTracer.objects.count()
-        pre_load_tracer_count = Tracer.objects.count()
-        pre_load_tracerlabel_count = TracerLabel.objects.count()
-
-        call_command(
-            "load_animals_and_samples",
-            animal_and_sample_table_filename=(
-                "DataRepo/example_data/obob_fasted_ace_glycerol_3hb_citrate_eaa_fa_multiple_tracers/"
-                "animal_sample_table.xlsx"
-            ),
-            skip_researcher_check=True,
-        )
-
-        post_load_sample_count = Sample.objects.count()
-        post_load_infusate_count = Infusate.objects.count()
-        post_load_infusatetracer_count = InfusateTracer.objects.count()
-        post_load_tracer_count = Tracer.objects.count()
-        post_load_tracerlabel_count = TracerLabel.objects.count()
-
-        self.assertEqual(num_samples, post_load_sample_count - pre_load_sample_count)
-        self.assertEqual(
-            num_infusates, post_load_infusate_count - pre_load_infusate_count
-        )
-        self.assertEqual(
-            num_infusatetracers,
-            post_load_infusatetracer_count - pre_load_infusatetracer_count,
-        )
-        self.assertEqual(num_tracers, post_load_tracer_count - pre_load_tracer_count)
-        self.assertEqual(
-            num_tracerlabels, post_load_tracerlabel_count - pre_load_tracerlabel_count
-        )
-
-    def test_multitracer_isocorr_load_1(self):
-        self.load_multitracer_data()
-        pre_load_group_count = PeakGroup.objects.count()
-        call_command(
-            "load_accucor_msruns",
-            accucor_file="DataRepo/example_data/obob_fasted_ace_glycerol_3hb_citrate_eaa_fa_multiple_tracers/"
-            "6eaafasted1_cor.xlsx",
-            protocol="Default",
-            date="2021-04-29",
-            researcher="Xianfeng Zeng",
-            new_researcher=False,
-            isocorr_format=True,
-        )
-        post_load_group_count = PeakGroup.objects.count()
-        # The number of samples in the isocorr xlsx file (not the samples file)
-        SAMPLES_COUNT = 30
-        PEAKDATA_ROWS = 86
-        PARENT_REC_COUNT = 15
-
-        self.assertEqual(
-            post_load_group_count - pre_load_group_count,
-            PARENT_REC_COUNT * SAMPLES_COUNT,
-            msg=f"PeakGroup record count should be the number of C12 PARENT lines [{PARENT_REC_COUNT}] times the "
-            f"number of samples [{SAMPLES_COUNT}] = [{PARENT_REC_COUNT * SAMPLES_COUNT}].",
-        )
-        self.assertEqual(
-            PeakData.objects.count(),
-            PEAKDATA_ROWS * SAMPLES_COUNT,
-            msg=f"PeakData record count should be the number of peakdata rows [{PEAKDATA_ROWS}] times the number of "
-            f"samples [{SAMPLES_COUNT}] = [{PEAKDATA_ROWS * SAMPLES_COUNT}].",
-        )
-
-    def test_multitracer_isocorr_load_2(self):
-        self.load_multitracer_data()
-        pre_load_group_count = PeakGroup.objects.count()
-        call_command(
-            "load_accucor_msruns",
-            accucor_file="DataRepo/example_data/obob_fasted_ace_glycerol_3hb_citrate_eaa_fa_multiple_tracers/"
-            "6eaafasted2_cor.xlsx",
-            protocol="Default",
-            date="2021-04-29",
-            researcher="Xianfeng Zeng",
-            new_researcher=False,
-            isocorr_format=True,
-        )
-        post_load_group_count = PeakGroup.objects.count()
-        # The number of samples in the isocorr xlsx file (not the samples file)
-        SAMPLES_COUNT = 30
-        PEAKDATA_ROWS = 81
-        PARENT_REC_COUNT = 15
-
-        self.assertEqual(
-            post_load_group_count - pre_load_group_count,
-            PARENT_REC_COUNT * SAMPLES_COUNT,
-            msg=f"PeakGroup record count should be the number of C12 PARENT lines [{PARENT_REC_COUNT}] times the "
-            f"number of samples [{SAMPLES_COUNT}] = [{PARENT_REC_COUNT * SAMPLES_COUNT}].",
-        )
-        self.assertEqual(
-            PeakData.objects.count(),
-            PEAKDATA_ROWS * SAMPLES_COUNT,
-            msg=f"PeakData record count should be the number of peakdata rows [{PEAKDATA_ROWS}] times the number of "
-            f"samples [{SAMPLES_COUNT}] = [{PEAKDATA_ROWS * SAMPLES_COUNT}].",
-        )
-
-    def test_multitracer_isocorr_load_3(self):
-        self.load_multitracer_data()
-        pre_load_group_count = PeakGroup.objects.count()
-        call_command(
-            "load_accucor_msruns",
-            accucor_file="DataRepo/example_data/obob_fasted_ace_glycerol_3hb_citrate_eaa_fa_multiple_tracers/"
-            "bcaafasted_cor.xlsx",
-            protocol="Default",
-            date="2021-04-29",
-            researcher="Xianfeng Zeng",
-            new_researcher=False,
-            isocorr_format=True,
-        )
-        post_load_group_count = PeakGroup.objects.count()
-        # The number of samples in the isocorr xlsx file (not the samples file)
-        SAMPLES_COUNT = 60
-        PEAKDATA_ROWS = 143
-        PARENT_REC_COUNT = 20
-
-        self.assertEqual(
-            post_load_group_count - pre_load_group_count,
-            PARENT_REC_COUNT * SAMPLES_COUNT,
-            msg=f"PeakGroup record count should be the number of C12 PARENT lines [{PARENT_REC_COUNT}] times the "
-            f"number of samples [{SAMPLES_COUNT}] = [{PARENT_REC_COUNT * SAMPLES_COUNT}].",
-        )
-        self.assertEqual(
-            PeakData.objects.count(),
-            PEAKDATA_ROWS * SAMPLES_COUNT,
-            msg=f"PeakData record count should be the number of peakdata rows [{PEAKDATA_ROWS}] times the number of "
-            f"samples [{SAMPLES_COUNT}] = [{PEAKDATA_ROWS * SAMPLES_COUNT}].",
-        )
-
-    def test_multilabel_sample_table_load(self):
-        num_samples = 156
-        num_infusates = 2
-        num_infusatetracers = 2
-        num_tracers = 2
-        num_tracerlabels = 4  # TracerLabel records are not unique. Note there would be 3 unique label records
-
-        pre_load_sample_count = Sample.objects.count()
-        pre_load_infusate_count = Infusate.objects.count()
-        pre_load_infusatetracer_count = InfusateTracer.objects.count()
-        pre_load_tracer_count = Tracer.objects.count()
-        pre_load_tracerlabel_count = TracerLabel.objects.count()
-
-        call_command(
-            "load_animals_and_samples",
-            animal_and_sample_table_filename=(
-                "DataRepo/example_data/obob_fasted_glc_lac_gln_ala_multiple_labels/animal_sample_table.xlsx"
-            ),
-            skip_researcher_check=True,
-        )
-
-        post_load_sample_count = Sample.objects.count()
-        post_load_infusate_count = Infusate.objects.count()
-        post_load_infusatetracer_count = InfusateTracer.objects.count()
-        post_load_tracer_count = Tracer.objects.count()
-        post_load_tracerlabel_count = TracerLabel.objects.count()
-
-        self.assertEqual(num_samples, post_load_sample_count - pre_load_sample_count)
-        self.assertEqual(
-            num_infusates, post_load_infusate_count - pre_load_infusate_count
-        )
-        self.assertEqual(
-            num_infusatetracers,
-            post_load_infusatetracer_count - pre_load_infusatetracer_count,
-        )
-        self.assertEqual(num_tracers, post_load_tracer_count - pre_load_tracer_count)
-        self.assertEqual(
-            num_tracerlabels, post_load_tracerlabel_count - pre_load_tracerlabel_count
-        )
-
-    def test_multilabel_isocorr_load_1(self):
-        self.load_multilabel_data()
-        pre_load_group_count = PeakGroup.objects.count()
-        call_command(
-            "load_accucor_msruns",
-            accucor_file="DataRepo/example_data/obob_fasted_glc_lac_gln_ala_multiple_labels/"
-            "alafasted_cor.xlsx",
-            protocol="Default",
-            date="2021-04-29",
-            researcher="Xianfeng Zeng",
-            new_researcher=False,
-            isocorr_format=True,
-        )
-        post_load_group_count = PeakGroup.objects.count()
-        # The number of samples in the isocorr xlsx file (not the samples file)
-        SAMPLES_COUNT = 84
-        PEAKDATA_ROWS = 94
-        PARENT_REC_COUNT = 13
-
-        self.assertEqual(
-            post_load_group_count - pre_load_group_count,
-            PARENT_REC_COUNT * SAMPLES_COUNT,
-            msg=f"PeakGroup record count should be the number of C12 PARENT lines [{PARENT_REC_COUNT}] times the "
-            f"number of samples [{SAMPLES_COUNT}] = [{PARENT_REC_COUNT * SAMPLES_COUNT}].",
-        )
-        self.assertEqual(
-            PeakData.objects.count(),
-            PEAKDATA_ROWS * SAMPLES_COUNT,
-            msg=f"PeakData record count should be the number of peakdata rows [{PEAKDATA_ROWS}] times the number of "
-            f"samples [{SAMPLES_COUNT}] = [{PEAKDATA_ROWS * SAMPLES_COUNT}].",
-        )
-
-    def test_multilabel_isocorr_load_2(self):
-        self.load_multilabel_data()
-        pre_load_group_count = PeakGroup.objects.count()
-        call_command(
-            "load_accucor_msruns",
-            accucor_file="DataRepo/example_data/obob_fasted_glc_lac_gln_ala_multiple_labels/"
-            "glnfasted1_cor.xlsx",
-            protocol="Default",
-            date="2021-04-29",
-            researcher="Xianfeng Zeng",
-            new_researcher=False,
-            isocorr_format=True,
-        )
-        post_load_group_count = PeakGroup.objects.count()
-        # The number of samples in the isocorr xlsx file (not the samples file)
-        SAMPLES_COUNT = 36
-        PEAKDATA_ROWS = 95
-        PARENT_REC_COUNT = 13
-
-        self.assertEqual(
-            post_load_group_count - pre_load_group_count,
-            PARENT_REC_COUNT * SAMPLES_COUNT,
-            msg=f"PeakGroup record count should be the number of C12 PARENT lines [{PARENT_REC_COUNT}] times the "
-            f"number of samples [{SAMPLES_COUNT}] = [{PARENT_REC_COUNT * SAMPLES_COUNT}].",
-        )
-        self.assertEqual(
-            PeakData.objects.count(),
-            PEAKDATA_ROWS * SAMPLES_COUNT,
-            msg=f"PeakData record count should be the number of peakdata rows [{PEAKDATA_ROWS}] times the number of "
-            f"samples [{SAMPLES_COUNT}] = [{PEAKDATA_ROWS * SAMPLES_COUNT}].",
-        )
-
-    def test_multilabel_isocorr_load_3(self):
-        self.load_multilabel_data()
-        pre_load_group_count = PeakGroup.objects.count()
-        call_command(
-            "load_accucor_msruns",
-            accucor_file="DataRepo/example_data/obob_fasted_glc_lac_gln_ala_multiple_labels/"
-            "glnfasted2_cor.xlsx",
-            protocol="Default",
-            date="2021-04-29",
-            researcher="Xianfeng Zeng",
-            new_researcher=False,
-            isocorr_format=True,
-            skip_samples=("bk",),
-        )
-        post_load_group_count = PeakGroup.objects.count()
-        # The number of samples in the isocorr xlsx file (not the samples file)
-        SAMPLES_COUNT = 36
-        PEAKDATA_ROWS = 95
-        PARENT_REC_COUNT = 13
-
-        self.assertEqual(
-            post_load_group_count - pre_load_group_count,
-            PARENT_REC_COUNT * SAMPLES_COUNT,
-            msg=f"PeakGroup record count should be the number of C12 PARENT lines [{PARENT_REC_COUNT}] times the "
-            f"number of samples [{SAMPLES_COUNT}] = [{PARENT_REC_COUNT * SAMPLES_COUNT}].",
-        )
-        self.assertEqual(
-            PeakData.objects.count(),
-            PEAKDATA_ROWS * SAMPLES_COUNT,
-            msg=f"PeakData record count should be the number of peakdata rows [{PEAKDATA_ROWS}] times the number of "
-            f"samples [{SAMPLES_COUNT}] = [{PEAKDATA_ROWS * SAMPLES_COUNT}].",
-        )
-
-    def test_labeled_elements_common_with_compound(self):
-        """
-        Test to ensure count 0 entries are not created when measured compound doesn't have that element
-        """
-        self.load_multilabel_data()
-        call_command(
-            "load_accucor_msruns",
-            accucor_file="DataRepo/example_data/obob_fasted_glc_lac_gln_ala_multiple_labels/"
-            "alafasted_cor.xlsx",
-            protocol="Default",
-            date="2021-04-29",
-            researcher="Xianfeng Zeng",
-            new_researcher=False,
-            isocorr_format=True,
-        )
-        pg = (
-            PeakGroup.objects.filter(msrun__sample__name="xzl5_panc")
-            .filter(name__exact="serine")
-            .filter(
-                peak_group_set__filename="alafasted_cor.xlsx",
-            )
-            .order_by("id", "peak_data__labels__element")
-            .distinct("id", "peak_data__labels__element")
-        )
-
-        self.assertEqual(pg.count(), 2)
-        self.assertEqual(pg.filter(peak_data__labels__element__exact="C").count(), 1)
-        self.assertEqual(pg.filter(peak_data__labels__element__exact="N").count(), 1)
 
 
 @override_settings(CACHES=settings.TEST_CACHES)
