+++ DataRepo/utils/__init__.py
@@ -11,6 +11,7 @@ from DataRepo.utils.exceptions import (  # AmbiguousCompoundDefinitionError,
     DryRun,
     DupeCompoundIsotopeCombos,
     DuplicateValues,
+    ExistingMSRun,
     HeaderConfigError,
     HeaderError,
     LoadingError,
@@ -41,6 +42,7 @@ __all__ = [
     "ConflictingValueError",
     "DupeCompoundIsotopeCombos",
     "DuplicateValues",
+    "ExistingMSRun",
     "HeaderError",
     "HeaderConfigError",
     "IsotopeObservationData",
